\name{concProb-package}
\alias{concProb-package}
\docType{package}
\title{
  Concordance Probability for large datasets
}
\description{
  This package gives two estimation methods to calculate the concordance probability in an efficient way, both for a continuous and a discrete setting.
}
\details{
  \tabular{ll}{
    Package: \tab concProb\cr
    Type: \tab Package\cr
    Version: \tab 1.0.0\cr
    License: \tab GPL (>= 2)\cr
  }
}

\references{arXiv:2105.10392}

\author{
  Ponnet J., Van Oirbeek R., Baesens B., Verdonck T.\cr
  Maintainer: Ponnet J. <jolien.ponnet@kuleuven.be>
}

